requirements.txt
fastapi
uvicorn
boto3
pydantic[email]
pytest-asyncio[httpx]
redis[aioredis]
httpx
PyJWT
python-multipart
